Week5/requirements.txt
PyMuPDF==1.23.8
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
tqdm==4.65.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
ipywidgets==8.1.1
jupyter==1.0.0
huggingface-hub==0.15.1
torch==2.0.1
transformers==4.30.2
rank-bm25==0.2.2
